added a fourth component to the BAP score based on B01.
